migrate `generate_smart_stamp_hash`
